Add gnome installation and data
Refactor files and add gnome installation and simple setting
configuration data.

Before this commit, the primary emphasis of this repository centered
around the installation and customization of Arch Linux KDE. However,
recent updates to KDE introduced disruptions to certain aspects of the
operating system and its customization. Recognizing this, I deemed it
necessary to explore an alternative desktop environment, such as GNOME.
Consequently, I have authored a guide for setting up and customizing
GNOME as a viable alternative.